AR_LSTM

+++ libfranka_ws/src/Model_based_Reinforcement_Learning_In_Teleoperation/Model_based_Reinforcement_Learning_In_Teleoperation/rl_agent_autoregressive/autoregressive_LSTM.py
@@ -38,7 +38,6 @@ from Model_based_Reinforcement_Learning_In_Teleoperation.rl_agent_autoregressive
 import Model_based_Reinforcement_Learning_In_Teleoperation.config.robot_config as cfg
 
 
-# --- Helper Function for Environment Creation ---
 def make_env_factory(rank: int, config: ExperimentConfig, traj_type: TrajectoryType, seed: int, randomize: bool, render_mode: Optional[str] = None):
     def _init():
         return TeleoperationEnvWithDelay(
@@ -58,7 +57,7 @@ class SequenceReplayBuffer:
     Target Shape: (Batch, Horizon, Output_Dim)
     """
     def __init__(self, buffer_size: int, device: torch.device):
-        # [FIX] Explicitly cast buffer_size to int to prevent numpy TypeError
+        # [FIX] Explicit int cast to handle float config values
         self.max_size = int(buffer_size)
         self.device = device
         self.ptr = 0
@@ -67,8 +66,7 @@ class SequenceReplayBuffer:
         self.seq_length = int(cfg.RNN_SEQUENCE_LENGTH)
         self.input_dim = int(cfg.ESTIMATOR_STATE_DIM)
         self.output_dim = int(cfg.N_JOINTS * 2)
-        # [FIX] Explicitly cast to int
-        self.ar_horizon = int(cfg.MAX_AR_STEPS) 
+        self.ar_horizon = int(cfg.MAX_AR_STEPS)
 
         self.input_sequences = np.zeros((self.max_size, self.seq_length, self.input_dim), dtype=np.float32)
         self.target_sequences = np.zeros((self.max_size, self.ar_horizon, self.output_dim), dtype=np.float32)
@@ -108,10 +106,10 @@ class LSTMTrainer:
         self.logger = self._setup_logging()
         self.tb_writer = SummaryWriter(log_dir=os.path.join(self.output_dir, "tensorboard"))
         
-        self.model = StateEstimator(output_dim=cfg.N_JOINTS * 2).to(self.device)
+        self.model = StateEstimator(output_dim=int(cfg.N_JOINTS * 2)).to(self.device)
         self.optimizer = torch.optim.Adam(self.model.parameters(), lr=cfg.ESTIMATOR_LEARNING_RATE)
         
-        # [FIX] Cast buffer size here as well to be safe
+        # [FIX] Explicit int cast for buffer size
         self.buffer = SequenceReplayBuffer(int(cfg.ESTIMATOR_BUFFER_SIZE), self.device)
         self.train_env, self.val_env = self._setup_environments()
         
@@ -142,15 +140,19 @@ class LSTMTrainer:
 
     def _collect_rollouts(self, env: DummyVecEnv) -> Tuple[np.ndarray, np.ndarray]:
         # Collect Input Sequence
-        delayed_flat_list = env.env_method("get_delayed_target_buffer", cfg.RNN_SEQUENCE_LENGTH)
+        # [FIX] Explicit int cast to resolve TypeError inside env_method
+        delayed_flat_list = env.env_method("get_delayed_target_buffer", int(cfg.RNN_SEQUENCE_LENGTH))
+        
         # Collect Target Sequence (Horizon = MAX_AR_STEPS)
-        ar_targets_flat = env.env_method("get_future_target_sequence", cfg.MAX_AR_STEPS)
+        # [FIX] Explicit int cast to resolve TypeError inside env_method
+        ar_targets_flat = env.env_method("get_future_target_sequence", int(cfg.MAX_AR_STEPS))
         
-        input_dim = cfg.ESTIMATOR_STATE_DIM
-        output_dim = cfg.N_JOINTS * 2
+        input_dim = int(cfg.ESTIMATOR_STATE_DIM)
+        output_dim = int(cfg.N_JOINTS * 2)
         
-        raw_inputs = np.array([buf.reshape(cfg.RNN_SEQUENCE_LENGTH, input_dim) for buf in delayed_flat_list])
-        raw_targets = np.array([buf.reshape(cfg.MAX_AR_STEPS, output_dim) for buf in ar_targets_flat])
+        # Reshape using int dimensions
+        raw_inputs = np.array([buf.reshape(int(cfg.RNN_SEQUENCE_LENGTH), input_dim) for buf in delayed_flat_list])
+        raw_targets = np.array([buf.reshape(int(cfg.MAX_AR_STEPS), output_dim) for buf in ar_targets_flat])
         
         valid_indices = []
         for i in range(len(raw_inputs)):
